+++ build.gradle
@@ -24,7 +24,6 @@ dependencies {
     runtimeOnly 'com.h2database:h2'
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
     testImplementation 'io.rest-assured:rest-assured:5.3.1'
-    implementation 'org.springframework.boot:spring-boot-starter-validation'
 }
 
 test {
