+++ src/test/java/roomescape/domain/ReservationTest.java
@@ -14,50 +14,6 @@ class ReservationTest {
 
     private final Clock clock = Clock.fixed(Instant.parse("2025-04-20T10:00:00Z"), ZoneId.systemDefault());
 
-    @DisplayName("예약자 이름은 null이면 예외를 발생한다.")
-    @Test
-    void validateNameNullThrowExceptionTest() {
-
-        // given
-        final LocalDate date = LocalDate.of(2025, 4, 21);
-        final LocalTime time = LocalTime.of(10, 0);
-
-        // when & then
-        assertThatThrownBy(() -> new Reservation(1L, null, date, time, clock))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("이름은 비어있을 수 없습니다.");
-    }
-
-    @DisplayName("예약자 이름이 비어있으면 예외를 발생한다.")
-    @Test
-    void validateNameBlankThrowExceptionTest() {
-
-        // given
-        final String name = "";
-        final LocalDate date = LocalDate.of(2025, 4, 21);
-        final LocalTime time = LocalTime.of(10, 0);
-
-        // when & then
-        assertThatThrownBy(() -> new Reservation(1L, name, date, time, clock))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("이름은 비어있을 수 없습니다.");
-    }
-
-    @DisplayName("예약자 이름이 띄어쓰기면 예외를 발생한다.")
-    @Test
-    void validateNameSpacingThrowExceptionTest() {
-
-        // given
-        final String name = " ";
-        final LocalDate date = LocalDate.of(2025, 4, 21);
-        final LocalTime time = LocalTime.of(10, 0);
-
-        // when & then
-        assertThatThrownBy(() -> new Reservation(1L, name, date, time, clock))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("이름은 비어있을 수 없습니다.");
-    }
-
     @DisplayName("예약 날짜가 이전 날짜면 예외를 발생한다.")
     @Test
     void validateDatePreviousThrowExceptionTest() {
@@ -81,7 +37,7 @@ class ReservationTest {
         final String name = "체체";
         final LocalDate date = LocalDate.of(2025, 4, 20);
         final LocalTime time = LocalTime.of(9, 59);
-        
+
         // when & then
         assertThatThrownBy(() -> new Reservation(1L, name, date, time, clock))
                 .isInstanceOf(IllegalArgumentException.class)
